+++ src/types/index.ts
@@ -1,5 +1,5 @@
-export type { RouteConfig } from './RouteConfig';
-export { RouteError, NotFoundError, OutletMissingError, ContentLoadError, ContentRenderError } from './RouteError';
-export { RouteBeginEvent, RouteDoneEvent, RouteErrorEvent } from './RouteEvent';
-export type { RouteInfo } from './RouteInfo';
-export type { RouterConfig } from './RouterConfig';
+export * from './RouteConfig';
+export * from './RouteContext';
+export * from './RouteError';
+export * from './RouteEvent';
+export * from './RouterConfig';